requirements.txt
streamlit
typing-extensions
langchain
openai
pydantic
python-dotenv
watchdog
wikipedia
chromadb
tiktoken